+++ astropy/cosmology/funcs.py
@@ -2,7 +2,139 @@
 """
 Convenience functions for `astropy.cosmology`.
 """
+import warnings
+import numpy as np
 from .core import get_current as _get_current
+from .core import CosmologyError
+from ..units import Quantity
+
+__doctest_requires__ = {'*': ['scipy.integrate']}
+
+def z_at_value(func, fval, zmin=0, zmax=1000, ztol=1e-5, maxfun=500):
+    """ Find the redshift `z` at which `func(z) = fval`.
+
+    This finds the redshift at which one of the cosmology functions or
+    methods (for example Planck13.distmod) is equal to a known value.
+
+    .. warning::
+      Make sure you understand the behaviour of the function that you
+      are trying to invert! Depending on the cosmology, there may not
+      be a unique solution. For example, in the standard Lambda CDM
+      cosmology, there are two redshifts which give an angular
+      diameter distance of 1500 Mpc, z ~ 0.7 and z ~ 3.8. To force
+      `z_at_value` to find the solution you are interested in, use the
+      `zmin` and `zmax` keywords to limit the search range (see the
+      example below).
+
+    Parameters
+    ----------
+    func : function or method
+       A function that takes a redshift as input.
+    fval : astropy.Quantity instance
+       The value of `func(z)`.
+    zmin : float, optional
+       The lower search limit for `z` (default 0).
+    zmax : float, optional
+       The upper search limit for `z` (default 1000).
+    ztol : float, optional
+       The relative error in `z` acceptable for convergence.
+    maxfun : int, optional
+       The maximum number of function evaluations allowed in the
+       optimization routine (default 500).
+
+    Returns
+    -------
+    z : float
+      The redshift `z` satisfying `zmin < z < zmax` and `func(z) =
+      fval` within `ztol`.
+
+    Notes
+    -----
+    This works for any arbitrary input cosmology, but is inefficient
+    if you want to invert a large number of values for the same
+    cosmology. In this case, it is faster to instead generate an array
+    of values at many closely-spaced redshifts that cover the relevant
+    redshift range, and then use interpolation to find the redshift at
+    each value you're interested in. For example, to efficiently find
+    the redshifts corresponding to 10^6 values of the distance modulus
+    in a Planck13 cosmology, you could do the following:
+
+    >>> import astropy.units as u
+    >>> from astropy.cosmology import Planck13, z_at_value
+
+    # Generate 10^6 distance moduli between 23 and 43 for which we
+    # want to find the corresponding redshifts
+    >>> Dvals = (23 + np.random.rand(1e6) * 20) * u.mag
+
+    # Make a grid of distance moduli covering the redshift range we
+    # need using 50 equally log-spaced values between zmin and
+    # zmax. We use log spacing to adequately sample the steep part of
+    # the curve at low distance moduli.
+    >>> zmin = z_at_value(Planck13.distmod, Dvals.min())
+    >>> zmax = z_at_value(Planck13.distmod, Dvals.max())
+    >>> zgrid = np.logspace(zmin, zmax)
+    >>> Dgrid = Planck13.distmod(zgrid)
+
+    # Finally interpolate to find the redshift at each distance modulus.
+    >>> zvals = np.interp(Dvals.value, zgrid, Dgrid.value)
+
+    Examples
+    --------
+    >>> import astropy.units as u
+    >>> from astropy.cosmology import Planck13, z_at_value
+
+    The age and lookback time are monotonic with redshift, and so a
+    unique solution can be found:
+
+    >>> z_at_value(Planck13.age, 2 * u.Gyr)
+    3.1981191749374629
+
+    The angular diameter is not monotonic however, and there are two
+    redshifts that give a value of 1500 Mpc. Use the zmin and zmax keywords
+    to find the one you're interested in:
+
+    >>> z_at_value(Planck13.angular_diameter_distance, 1500 * u.Mpc, zmax=1.5)
+    0.68127769625288614
+    >>> z_at_value(Planck13.angular_diameter_distance, 1500 * u.Mpc, zmin=2.5)
+    3.7914918534022011
+
+    Also note that the luminosity distance and distance modulus (two
+    other commonly inverted quantities) are monotonic in flat and open
+    universes, but not in closed universes.
+    """
+    from scipy.optimize import fminbound
+
+    fval_zmin = func(zmin)
+    fval_zmax = func(zmax)
+    if np.sign(fval - fval_zmin) != np.sign(fval_zmax - fval):
+        warnings.warn("""\
+fval is not bracketed by func(zmin) and func(zmax). This means either
+there is no solution, or that there is more than one solution between
+zmin and zmax satisfying fval = func(z).""")
+
+    if isinstance(fval_zmin, Quantity):
+        unit = fval_zmin.unit
+        val = fval.to(unit).value
+        f = lambda z: abs(func(z).value - val)
+    else:
+        f = lambda z: abs(func(z) - fval)
+
+    zbest, resval, ierr, ncall = fminbound(f, zmin, zmax, maxfun=maxfun,
+                                           full_output=1)
+
+    if ierr != 0:
+        warnings.warn('Maximum number of function calls ({}) reached'.format(
+            ncall))
+
+    if np.allclose(zbest, zmax):
+        raise CosmologyError("Best guess z is very close the upper z limit.\n"
+                             "Try re-running with a different zmax.")
+    elif np.allclose(zbest, zmin):
+        raise CosmologyError("Best guess z is very close the lower z limit.\n"
+                             "Try re-running with a different zmin.")
+
+    return zbest
+
 
 def age(z, cosmo=None):
     """ Age of the universe in Gyr at redshift `z`.
@@ -16,7 +148,11 @@ def age(z, cosmo=None):
     -------
     t : astropy.units.Quantity
       The age of the universe in Gyr at each input redshift.
-        """
+
+    See Also
+    --------
+    z_at_value : Find the redshift corresponding to an age.
+    """
     if cosmo is None:
         cosmo = _get_current()
     return cosmo.age(z)
@@ -140,6 +276,10 @@ def distmod(z, cosmo=None):
     -------
     distmod : astropy.units.Quantity
       Distance modulus at each input redshift.
+
+    See Also
+    --------
+    z_at_value : Find the redshift corresponding to a distance modulus.
     """
     if cosmo is None:
         cosmo = _get_current()
@@ -217,6 +357,10 @@ def lookback_time(z, cosmo=None):
     -------
     t : astropy.units.Quantity
       Lookback time at each input redshift.
+
+    See Also
+    --------
+    z_at_value : Find the redshift corresponding to a lookback time.
     """
     if cosmo is None:
         cosmo = _get_current()
@@ -281,6 +425,10 @@ def luminosity_distance(z, cosmo=None):
     -------
     lumdist : astropy.units.Quantity
       Angular diameter distance at each input redshift.
+
+    See Also
+    --------
+    z_at_value : Find the redshift corresponding to a luminosity distance.
     """
     if cosmo is None:
         cosmo = _get_current()