+++ CHANGES.rst
@@ -17,6 +17,9 @@ New Features
   - `age` and `comoving_volume` convenience functions have been added to
     `astropy.cosmology`. [#1902]
 
+  - Added `astropy.cosmology.z_at_value` to find the redshift at which a
+    cosmology function matches a desired value. [#1909]
+
 - ``astropy.io.ascii``
 
   - The ``astropy.io.ascii`` output formats ``latex`` and ``aastex`` accept a